requirements.txt
apipkg==1.4
attrs==18.1.0
boto3==1.7.26
botocore==1.10.26
certifi==2018.4.16
chardet==3.0.4
dnspython==1.15.0
docutils==0.14
execnet==1.5.0
idna==2.6
jmespath==0.9.3
more-itertools==4.1.0
numpy==1.14.3
pandas==0.23.0
pluggy==0.6.0
py==1.5.3
pymongo==3.6.1
pytest==3.5.1
pytest-forked==0.2
pytest-xdist==1.22.2
python-dateutil==2.7.3
python-dotenv==0.8.2
pytz==2018.4
redis==2.10.6
requests==2.18.4
s3transfer==0.1.13
six==1.11.0
toolz==0.9.0
urllib3==1.22
xlrd==1.1.0
